requirements.txt
flask
flask-cors
chromadb
sentence-transformers
huggingface_hub
pymongo
python-dotenv
pandas